API/requirement.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
commonmark==0.9.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
httpie==3.2.1
idna==3.4
multidict==6.0.2
oauthlib==3.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==12.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
